Separate out model modifier
This creates a separate interface to changing the model.  This makes it
easier to ensure that all changes to the model are performed in the
correct thread and at the correct time.  It also makes it a lot clearer
what the model’s behavior is and what the modifier’s behavior is.