Replace deleteLine with deleteLines
This allows us to use insertLines in a straightforward fashion for
deleteLines.  This model is so simple.  I believe scrollDisplayUp can do
without the linesChangedFrom as well, as the cursor, according to
terminfo(5), should always be at scrollingRegion.top() anyway.  This
actually means that scrollDisplayUp() is nothing but insertLines(1), so
perhaps we can simplify it to that.