Implement new parser
The new parser is implemented using static methods.  This is for speed
reasons only and I’m not sure how much this saves.  Will have to test
this a bit.  Using instances would be a lot simpler for many things, but
it would force us to keep track of the character set in some other way.