Rename ScrollDisplayUp to ReverseLineFeed
This patch also makes it consistent with how screen and xterm interpret
it when the cursor isn’t at the top of the scrolling region, which is to
move the cursor up one line.  This is consistent with VT100.
Terminfo(5), however, states that the “ri” capability, which is the one
linked to \E[M, is undefined if the cursor isn’t at the top of the
scrolling region, so programs using terminfo shouldn’t be doing any
cursor movement this way.